+++ client/components/body/Admin.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function Admin() {
+  return (
+    <div className="admin">
+      <h1>Admin!!</h1>
+    </div>
+  );
+}
+
+export default Admin;
