+++ client/components/body/LoggedIn.jsx
@@ -1,0 +1,11 @@
+import React from 'react';
+
+function LoggedIn() {
+  return (
+    <div className="userLoginPage">
+      <h1>Welcome back, user!!</h1>
+    </div>
+  );
+}
+
+export default LoggedIn;
