+++ client/index.js
@@ -1,12 +1,12 @@
 import React from 'react';
 import { render } from 'react-dom';
 import { BrowserRouter } from 'react-router-dom';
-// eslint-disable-next-line import/extensions
+
 import App from './App';
 
 render(
   <BrowserRouter>
     <App />
   </BrowserRouter>,
-  document.getElementById('app'),
+  document.getElementById('app')
 );
